nyagaks/requirements.txt
Django>=3.0,<3.1
wagtail>=2.9,<2.10
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.1
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9.2
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.9
